refactored redirect handler into separate file
